requirements.txt
pandas==0.20.2
tqdm==4.11.2
imgaug==0.2.5
mxnet==0.11.0
numpy==1.13.3
pytest==3.1.1
Pillow==5.0.0
begins==0.9
opencv-python
